+++ playwright-e2e/tests/dsm/mailing-list/rgp-mail-list-download.spec.ts-snapshots/requirement-list-chromium-linux.png
