+++ playwright-e2e/utils/file-utils.ts
@@ -8,12 +8,15 @@ export interface MailListCSV {
   lastName?: string;
 }
 
-export async function readMailListCSVFile(filePath: string): Promise<MailListCSV[]> {
+export async function readMailListCSVFile(filePath: string | null): Promise<MailListCSV[]> {
+  if (filePath == null) {
+    throw Error('filePath is null');
+  }
   const results: MailListCSV[] = [];
 
   return new Promise((resolve, reject) => {
     fs.createReadStream(filePath)
-    .pipe(csv())
+    .pipe(csv({ strict: true }))
     .on('data', (data) => results.push(data))
     .on('end', () => resolve(results))
     .on('error', (error) => reject(error));
